+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](#unreleased)
 ### Added
+- Added /users/{user_id}/teams API endpoint (UsersApi::getUserTeams())
+
 ### Changed
 - Refactored Exceptions structure and inheritances